added build-impl.xml~ to ignore moved control and info panel to left tried to set fixed sizes for both panels fixed organism body direction initialization (random was from 0.0 to 1.0 but not from 0.0 to 2*PI) fixed "organisms prefer upper left corner"-bug; bodies now use double coordinates to prevent loss of small movements on some axes

+++ src/jgpfun/world2d/Body2d.java
@@ -14,12 +14,12 @@ public class Body2d {
 
     public double dir;
 
-    public int x;
+    public double x;
 
-    public int y;
+    public double y;
 
 
-    public Body2d(int x, int y, double dir, FoodFinder foodFinder, WallSense wallSense) {
+    public Body2d(double x, double y, double dir, FoodFinder foodFinder, WallSense wallSense) {
         this.x = x;
         this.y = y;
         this.dir = dir;
